change file for commonJS
