+++ sh.exe.stackdump
@@ -1,0 +1,24 @@
+MSYS-1.0.12 Build:2012-07-05 14:56
+Exception: STATUS_ACCESS_VIOLATION at eip=6804050C
+eax=00000000 ebx=0022FD4C ecx=7FFDF000 edx=00000000 esi=00000AC4 edi=680A1174
+ebp=0022FD0C esp=0022FCE4 program=C:\Program Files\Git\bin\sh.exe
+cs=001B ds=0023 es=0023 fs=003B gs=0000 ss=0023
+Stack trace:
+Frame     Function  Args
+0022FD0C  6804050C  (0022FD4C, 00000000, 0022FD5C, 680549BB)
+0022FD6C  6805323B  (00000AC4, FFFFFFFF, 0022FD9C, 68004EE3)
+0022FD9C  68004F78  (0000003A, 680849C0, 0022FDCC, 68006ED3)
+0022FDBC  6800506F  (0000003A, 0A0152E0, 0022FDEC, 68084616)
+0022FDEC  68084642  (0000003A, 680A1368, 00000003, 00403BA5)
+0022FE0C  00401F4A  (0000003A, 00000000, 0022FE3C, 00401F2F)
+0022FE3C  00401F38  (0000003A, 00000000, 0022FE8C, 00401A61)
+0022FE8C  00401A6F  (00000003, 68573F1C, 0A010278, 00000000)
+0022FEE8  68004B08  (00000000, 00000000, 0022FF28, 0048BC00)
+0022FF08  68004C5F  (00401248, 00000000, 00000000, 00000000)
+0022FF28  68004C98  (00000000, 00000000, 00000000, 00000000)
+0022FF58  0048BB08  (00401248, 00000000, 00000000, 00000000)
+0022FF88  0040103D  (7FFD5000, 0022FFD4, 777037EB, 7FFD5000)
+0022FF94  771AED6C  (7FFD5000, 77F81EF1, 00000000, 00000000)
+0022FFD4  777037EB  (00401000, 7FFD5000, 00000000, 00000000)
+0022FFEC  777037BE  (00401000, 7FFD5000, 00000000, 78746341)
+End of stack trace (more stack frames may be present)
