+++ src/main.js
@@ -1,5 +1,8 @@
 'use strict';
 
+require('./check');
 require('./form');
-require('./game');
+require('./game/game');
+require('./getReviews/getReviews');
+require('./getReviews/getReviewBlock');
 require('./reviews');
